parser part1 done
the parser can take A and C instructions and convert to hack machine language. It still dont changing the other lines, and every line is stored in a list of strings. Theres a list of ranges tracking wich lines werent converted to machine language yet, and that can be usefull for the next steps of this excercise.